Change notifications dialog configuration Show dialog on client response

+++ android/BeautyApp/beautician-trunk/src/com/pictureit/noambaroz/beauticianapp/data/DataProvider.java
@@ -31,6 +31,10 @@ public class DataProvider extends ContentProvider {
 	public static final String COL_TREATMENT_TIME = "treatment_time";
 	public static final String COL_NEED_TO_SHOW_DIALOG = "is_alarm_fired";
 	public static final String COL_ADDRESS = "address";
+	public static final String COL_IS_PLAYED = "isplayed";
+
+	public static final String TABLE_CONFIRMED_TREATMENTS = "tableconfirmedtreatments";
+	public static final String COL_CUSTOMER_TELEPHONE = "telephone";
 
 	private DbHelper dbHelper;
 
@@ -38,11 +42,15 @@ public class DataProvider extends ContentProvider {
 			.parse("content://com.pictureit.noambaroz.beauticianapp.provider/" + TABLE_ORDERS_AROUND_ME);
 	public static final Uri CONTENT_URI_ALARMS = Uri.parse("content://com.pictureit.noambaroz.beauticianapp.provider/"
 			+ TABLE_ALARMS);
+	public static final Uri CONTENT_CONFIRMED_TREATMENTS = Uri
+			.parse("content://com.pictureit.noambaroz.beauticianapp.provider/" + TABLE_CONFIRMED_TREATMENTS);
 
 	private static final int ORDERS_AROUND_ME_ALLROWS = 7;
 	private static final int ORDERS_AROUND_ME_SINGLE_ROW = 8;
 	private static final int ALARMS_ALLROWS = 3;
 	private static final int ALARMS_SINGLE_ROW = 4;
+	private static final int CONFIRMED_TREATMENTS_ALLROWS = 5;
+	private static final int CONFIRMED_TREATMENTS_SINGLE_ROW = 6;
 
 	private static final UriMatcher uriMatcher;
 	static {
@@ -54,6 +62,11 @@ public class DataProvider extends ContentProvider {
 
 		uriMatcher.addURI("com.pictureit.noambaroz.beauticianapp.provider", TABLE_ALARMS, ALARMS_ALLROWS);
 		uriMatcher.addURI("com.pictureit.noambaroz.beauticianapp.provider", TABLE_ALARMS + "/#", ALARMS_SINGLE_ROW);
+
+		uriMatcher.addURI("com.pictureit.noambaroz.beauticianapp.provider", TABLE_CONFIRMED_TREATMENTS,
+				CONFIRMED_TREATMENTS_ALLROWS);
+		uriMatcher.addURI("com.pictureit.noambaroz.beauticianapp.provider", TABLE_CONFIRMED_TREATMENTS + "/#",
+				CONFIRMED_TREATMENTS_SINGLE_ROW);
 	}
 
 	@Override
@@ -70,11 +83,13 @@ public class DataProvider extends ContentProvider {
 		switch (uriMatcher.match(uri)) {
 		case ORDERS_AROUND_ME_ALLROWS:
 		case ALARMS_ALLROWS:
+		case CONFIRMED_TREATMENTS_ALLROWS:
 			qb.setTables(getTableName(uri));
 			break;
 
 		case ORDERS_AROUND_ME_SINGLE_ROW:
 		case ALARMS_SINGLE_ROW:
+		case CONFIRMED_TREATMENTS_SINGLE_ROW:
 			qb.setTables(getTableName(uri));
 			qb.appendWhere("_id = " + uri.getLastPathSegment());
 			break;
@@ -99,11 +114,11 @@ public class DataProvider extends ContentProvider {
 		long id;
 		switch (uriMatcher.match(uri)) {
 		case ORDERS_AROUND_ME_ALLROWS:
-			id = db.insertOrThrow(TABLE_ORDERS_AROUND_ME, null, values);
-			break;
 		case ALARMS_ALLROWS:
-			id = db.insertOrThrow(TABLE_ALARMS, null, values);
+		case CONFIRMED_TREATMENTS_ALLROWS:
+			id = db.insertOrThrow(getTableName(uri), null, values);
 			break;
+
 		default:
 			throw new IllegalArgumentException("Unsupported URI: " + uri);
 		}
@@ -116,25 +131,25 @@ public class DataProvider extends ContentProvider {
 	@Override
 	public int delete(Uri uri, String selection, String[] selectionArgs) {
 		SQLiteDatabase db = dbHelper.getWritableDatabase();
-		db.beginTransaction();
+
 		int count;
 		switch (uriMatcher.match(uri)) {
 		case ORDERS_AROUND_ME_ALLROWS:
 		case ALARMS_ALLROWS:
+		case CONFIRMED_TREATMENTS_ALLROWS:
 			count = db.delete(getTableName(uri), selection, selectionArgs);
 			break;
 
 		case ORDERS_AROUND_ME_SINGLE_ROW:
 		case ALARMS_SINGLE_ROW:
+		case CONFIRMED_TREATMENTS_SINGLE_ROW:
 			count = db.delete(getTableName(uri), "_id = ?", new String[] { uri.getLastPathSegment() });
 			break;
 
 		default:
 			throw new IllegalArgumentException("Unsupported URI: " + uri);
 		}
-		db.setTransactionSuccessful();
-		db.endTransaction();
-		// getContext().getContentResolver().notifyChange(uri, null);
+		getContext().getContentResolver().notifyChange(uri, null);
 		return count;
 	}
 
@@ -146,11 +161,13 @@ public class DataProvider extends ContentProvider {
 		switch (uriMatcher.match(uri)) {
 		case ORDERS_AROUND_ME_ALLROWS:
 		case ALARMS_ALLROWS:
+		case CONFIRMED_TREATMENTS_ALLROWS:
 			count = db.update(getTableName(uri), values, selection, selectionArgs);
 			break;
 
 		case ORDERS_AROUND_ME_SINGLE_ROW:
 		case ALARMS_SINGLE_ROW:
+		case CONFIRMED_TREATMENTS_SINGLE_ROW:
 			count = db.update(getTableName(uri), values, "_id = ?", new String[] { uri.getLastPathSegment() });
 			break;
 
@@ -171,6 +188,9 @@ public class DataProvider extends ContentProvider {
 		case ALARMS_ALLROWS:
 		case ALARMS_SINGLE_ROW:
 			return TABLE_ALARMS;
+		case CONFIRMED_TREATMENTS_ALLROWS:
+		case CONFIRMED_TREATMENTS_SINGLE_ROW:
+			return TABLE_CONFIRMED_TREATMENTS;
 		}
 
 		return null;
